[KFI]refactor(pickers-react): move example app to storybook
