+++ examples/sn-react-component-docs/tslint.json
@@ -2,6 +2,7 @@
   "extends": ["../../tslint.json"],
   "rules": {
     "no-var-requires": false,
-    "completed-docs": false
+    "completed-docs": false,
+    "variable-name": false
   }
 }
