--- packages/sn-pickers-react/example/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-    <link rel="shortcut icon" href="./favicon.ico" />
-    <title>Pickers react example</title>
-  </head>
-  <style>
-    #app {
-      text-align: center;
-      color: #2c3e50;
-      max-width: 800px;
-      margin: 60px auto;
-    }
-  </style>
-
-  <body>
-    <noscript> You need to enable JavaScript to run this app. </noscript>
-    <div id="app">
-      <div id="root"></div>
-    </div>
-  </body>
-</html>
