--- packages/sn-pickers-react/example/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import App from './ExampleApp'
-
-ReactDOM.render(<App />, document.getElementById('root'))
